refactor: start splitting request messages
